gg p1

+++ index.php
@@ -1,4 +1,6 @@
 <?php
+session_start();
+
 require_once('database.php');
 require_once('user.php');
 require_once('db_pdo.php');
@@ -6,62 +8,63 @@ $config = require_once('config.php');
 
 use DB\DB_PDO as DB;
 
-session_start();
-
 $PDOConn = DB::getInstance($config);
 $conn = $PDOConn->getConnection(); //Mi connetto
 
 $userDTO = new UserDTO($conn);
 
-if ($_SERVER["REQUEST_METHOD"] == "POST") {
-    if (isset($_POST['firstname'])) {
-        $firstname = $_POST['firstname'];
-        $lastname = $_POST['lastname'];
-        $email = $_POST['email'];
-        $password = $_POST['password'];
-        $admin = $_POST['admin'];
+if ($_SESSION['userLogin']['admin'] == 1) {
+    if ($_SERVER["REQUEST_METHOD"] == "POST") {
+        if (isset($_POST['firstname'])) {
+            $firstname = $_POST['firstname'];
+            $lastname = $_POST['lastname'];
+            $email = $_POST['email'];
+            $password = $_POST['password'];
+            $admin = $_POST['admin'];
 
-        $res = $userDTO->saveUser([
-            'firstname' => $firstname,
-            'lastname' => $lastname,
-            'email' => $email,
-            'password' => $password,
-            'admin' => $admin
-        ]);
+            $res = $userDTO->saveUser([
+                'firstname' => $firstname,
+                'lastname' => $lastname,
+                'email' => $email,
+                'password' => $password,
+                'admin' => $admin
+            ]);
+        }
+
+        if (isset($_REQUEST['id']) && $_REQUEST['action'] == 'update') {
+            $id = intval($_REQUEST['id']);
+            $firstname = $_POST['firstnameUp'];
+            $lastname = $_POST['lastnameUp'];
+            $email = $_POST['emailUp'];
+            $password = $_POST['passwordUp'];
+            $admin = $_POST['adminUp'];
+
+            $res = $userDTO->updateUser([
+                'id' => $id,
+                'firstname' => $firstname,
+                'lastname' => $lastname,
+                'email' => $email,
+                'password' => $password,
+                'admin' => $admin
+            ]);
+        }
+    }
+    if (isset($_GET['id']) && $_GET['action'] == 'delete') {
+        $id = intval($_GET['id']);
+
+        $res = $userDTO->deleteUser($id);
+
+        header('Location: index.php');
+        exit;
     }
 
-    if (isset($_REQUEST['id']) && $_REQUEST['action'] == 'update') {
-        $id = intval($_REQUEST['id']);
-        $firstname = $_POST['firstnameUp'];
-        $lastname = $_POST['lastnameUp'];
-        $email = $_POST['emailUp'];
-        $password = $_POST['passwordUp'];
-        $admin = $_POST['adminUp'];
-
-        $res = $userDTO->updateUser([
-            'id' => $id,
-            'firstname' => $firstname,
-            'lastname' => $lastname,
-            'email' => $email,
-            'password' => $password,
-            'admin' => $admin
-        ]);
-    }
-
-}
-
-if (isset($_GET['id']) && $_GET['action'] == 'delete') {
-    $id = intval($_GET['id']);
-
-    $res = $userDTO->deleteUser($id);
-
-    header('Location: index.php');
+    $res = $userDTO->getAll();
+} else {
+    header('Location: notAdminIndex.php');
     exit;
 }
 
-var_dump($_SESSION['username']);
 
-$res = $userDTO->getAll();
 ?>
 
 
@@ -82,7 +85,7 @@ $res = $userDTO->getAll();
         <div class="container-fluid">
             <a class="navbar-brand">PS16</a>
             <form class="d-flex" role="search">
-                <button class="btn btn-outline-primary" type="submit">Logout</button>
+                <a href="logout.php" class="btn btn-outline-primary" type="submit">Logout</a>
             </form>
         </div>
     </nav>
@@ -130,14 +133,22 @@ $res = $userDTO->getAll();
                                 <?= $record["password"] ?>
                             </td>
                             <td>
-                                <?= $record["admin"] ?>
+                                <?php
+                                if ($record["admin"] == 0) {
+                                    echo "No";
+                                } else {
+                                    echo "Si";
+                                }
+                                ?>
                             </td>
                             <td>
                                 <a href="#" class="btn btn-primary" data-bs-toggle="modal"
                                     data-bs-target="#<?= $modalId ?>">Modifica</a>
                                 <a href="index.php?action=delete&id=<?= $record["id"] ?>" class="btn btn-danger">Elimina</a>
                             </td>
+
                         </tr>
+
 
                         <!-- Modale per la modifica -->
                         <div class="modal fade" id="<?= $modalId ?>" tabindex="-1" aria-labelledby="<?= $modalId ?>Label"
